requirements.txt
dotenv
PySide6>=6.7.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0